結合テストケースを作った
